Create an SSH Key
A SSH Key can authenticate clients useing a variety of different methods. The most basice of these is password authernication. Each key pair consists of a public key and a private key. The private key is retained by the client and should be kept secret. As an additional precaution, the key can be encrypted on disk with a passphrase. While the public key is uploaded to a remote server that you want to be log into with SSH.